fix: move hook for setup steps to bcsc theme hooks
